refactor(ai): extract AI initialization logic to shared utility
• Created new ai-init.util.ts to handle AI provider initialization
• Removes duplicate initialization code from branch and commit coordinators
• Improves maintainability by centralizing AI setup logic
• No functional changes
