+++ packages/gitguard/src/controllers/branch/branch.coordinator.ts
@@ -1,6 +1,4 @@
-import chalk from "chalk";
 import { BranchCommandOptions } from "../../commands/branch.js";
-import { AIFactory } from "../../services/factories/ai.factory.js";
 import { GitService } from "../../services/git.service.js";
 import { GitHubService } from "../../services/github.service.js";
 import { LoggerService } from "../../services/logger.service.js";
@@ -11,6 +9,7 @@ import { AIProvider } from "../../types/ai.types.js";
 import { PRAnalysisResult } from "../../types/analysis.types.js";
 import { Config, GitConfig } from "../../types/config.types.js";
 import { SecurityCheckResult } from "../../types/security.types.js";
+import { initializeAI } from "../../utils/ai-init.util.js";
 import { loadConfig } from "../../utils/config.util.js";
 import { BranchAIController } from "./branch-ai.controller.js";
 import { BranchAnalysisController } from "./branch-analysis.controller.js";
@@ -108,61 +107,8 @@ async function initializeServices({
   const github = new GitHubService({ config, logger, git });
   const security = new SecurityService({ config, logger });
 
-  logger.info("\n🔍 Checking AI configuration...");
-  let ai: AIProvider | undefined;
-
   const isAIRequested = options.ai ?? config.ai?.enabled;
-  if (isAIRequested) {
-    try {
-      if (!config.ai?.provider) {
-        // Create default fallback config
-        const fallbackConfig: Config = {
-          ...config,
-          ai: {
-            ...config.ai, // Preserve any existing AI config
-            enabled: true,
-            provider: "openai",
-            openai: {
-              model: "gpt-4-turbo",
-            },
-          },
-        };
-
-        ai = AIFactory.create({ config: fallbackConfig, logger });
-
-        logger.warn(
-          "\n⚠️  AI requested but no provider configured in settings",
-        );
-        logger.info(
-          "\n💡 Using default OpenAI configuration for offline prompts. To configure AI properly:",
-        );
-        logger.info(chalk.cyan("\n1. Run setup command:"));
-        logger.info(chalk.dim("   gitguard init"));
-        logger.info(chalk.cyan("\n2. Or manually update your config file:"));
-        logger.info(
-          chalk.dim("   .gitguard/config.json or ~/.gitguard/config.json"),
-        );
-        logger.info(
-          chalk.dim("\nTip: Run 'gitguard init --help' for more options"),
-        );
-      } else {
-        ai = AIFactory.create({ config, logger });
-      }
-
-      if (ai) {
-        logger.info(`✅ AI initialized using ${ai.getName()}`);
-      } else {
-        logger.warn(
-          `⚠️  AI configuration found but initialization failed. Falling back to offline prompts.`,
-        );
-      }
-    } catch (error) {
-      logger.warn("⚠️  Failed to initialize AI provider:", error);
-      logger.info("💡 Falling back to offline prompts");
-    }
-  } else {
-    logger.info("ℹ️  AI analysis disabled");
-  }
+  const ai = initializeAI({ config, logger, isAIRequested });
 
   const prService = new PRService({
     config,
